♻️ Refactor mapping
